requirements.txt
fastapi 
numpy 
uvicorn 
typing 
pymongo 
pydantic
psycopg2-binary